+++ src/services/ConsentGroupService.php
@@ -2,25 +2,34 @@
 
 namespace dutchheight\cookieboss\services;
 
+use Craft;
 use craft\base\Component;
+use craft\db\Query;
+use craft\events\ConfigEvent;
+use craft\helpers\Db;
+use craft\helpers\StringHelper;
+
 use dutchheight\cookieboss\records\ConsentGroup;
 
 class ConsentGroupService extends Component {
 
-    public static function getAll() {
+    const CONFIG_KEY = 'plugins.cookie-boss.consentGroups';
+    const TABLE = '{{%cookieboss_consentgroup}}';
+
+    public function getAll() {
         return ConsentGroup::find()->orderBy('id')->all();
     }
 
-    public static function getAllByHandle($handle) {
+    public function getAllByHandle($handle) {
         return ConsentGroup::find()->where(['handle' => $handle])->one();
     }
 
-    public static function getAllAsSelectOptions() {
+    public function getAllAsSelectOptions() {
         $returner[] = [
             'value' => null,
             'label' => '-'
         ];
-        foreach (ConsentGroupService::getAll() as $key => $value) {
+        foreach ($this->getAll() as $value) {
             $returner[] = [
                 'value' => $value['id'],
                 'label' => $value['name']
@@ -30,59 +39,128 @@ class ConsentGroupService extends Component {
         return $returner;
     }
 
-    public static function getAllEnabled() {
+    public function getAllEnabled() {
         return ConsentGroup::find()->where(['enabled' => 1])->orderBy('id')->all();
     }
 
-    public static function deleteAll() {
+    public function deleteAll() {
         return ConsentGroup::deleteAll();
     }
 
-    public static function selectId($el) {
+    public function selectId($el) {
         if ($el['id']) {
             return (int)$el['id'];
         }
     }
 
-    public static function updateAll(Array $groups) {
-
+    public function updateAll(Array $groups) {
         // All id's that are left over should be deleted
-        $currentIds = ConsentGroup::find()->select(['id'])->asArray()->all();
-        $currentIds = array_map('self::selectId', $currentIds);
+        $currentIds = array_map(
+            'self::selectId',
+            ConsentGroup::find()->select(['id'])->asArray()->all()
+        );
 
         foreach ($groups as $group) {
-            if ($group["enabled"] == "") {
-                continue;
-            }
+            // Check if cg is new
             if ($group['id']) {
                 // Delete from array because id is still present
                 if (($key = array_search($group['id'], $currentIds)) !== false) {
                     unset($currentIds[$key]);
                 }
+
+                // Check if cg has UUID
                 $cg = ConsentGroup::findOne(['id' => $group["id"]]);
+                if (!$cg->uid) {
+                    $cg->uid = Db::uidById(self::TABLE, $cg->id);
+                }
             } else {
+                // Make UUID for new Group
                 $cg = new ConsentGroup();
+                $cg->uid = StringHelper::UUID();
             }
 
-            $cg->enabled = (!empty($group["enabled"]) ? $group["enabled"] : 0);
+            // Update props
+            $cg->enabled      = (!empty($group["enabled"]) ? $group["enabled"] : 0);
             $cg->defaultValue = (!empty($group["defaultValue"]) ? $group["defaultValue"] : 0);
-            $cg->required = (!empty($group["required"]) ? $group["required"] : 0);
+            $cg->required     = (!empty($group["required"]) ? $group["required"] : 0);
+            $cg->handle       = $group["handle"];
+            $cg->name         = $group["name"];
+            $cg->desc         = $group["desc"];
 
-            $cg->handle = $group["handle"];
-            $cg->name = $group["name"];
-            $cg->desc = $group["desc"];
-
-            if (!$cg->validate()) {
-                throw new \Exception(json_encode($cg->errors));
-            }
-            $cg->save();
+            // Save to config
+            $this->saveConsentGroup($cg);
         }
 
         if (count($currentIds) > 0) {
-            $deleteIds = implode(', ', $currentIds);
-            ConsentGroup::deleteAll('id IN (' . $deleteIds . ')');
+            $uids = Db::uidsByIds(self::TABLE, $currentIds);
+            foreach ($uids as $uid) {
+                Craft::$app->projectConfig->remove(self::CONFIG_KEY . "." .$uid);
+            }
         }
 
         return true;
+    }
+
+    private function saveConsentGroup(ConsentGroup $consentGroup) {
+
+        if (!$consentGroup->validate()) {
+            return false;
+        }
+
+        // Save it to the project config
+        Craft::$app->projectConfig->set(self::CONFIG_KEY . "." .$consentGroup->uid, [
+            'enabled'      => $consentGroup->enabled,
+            'defaultValue' => $consentGroup->defaultValue,
+            'required'     => $consentGroup->required,
+            'handle'       => $consentGroup->handle,
+            'name'         => $consentGroup->name,
+            'desc'         => $consentGroup->desc
+        ]);
+
+        return true;
+    }
+
+    //
+    // ─── Project config event handlers ───────────────────────────────────────────────────────────────────────────
+    //
+
+    public function handleChangedConsentGroup(ConfigEvent $event)
+    {
+        $uid = $event->tokenMatches[0];
+
+        $id = (new Query())
+            ->select(['id'])
+            ->from(self::TABLE)
+            ->where(['uid' => $uid])
+            ->scalar();
+
+        $isNew = empty($id);
+        if ($isNew) {
+            Craft::$app->db->createCommand()->insert(self::TABLE, [
+                    'uid'          => $uid,
+                    'enabled'      => $event->newValue['enabled'],
+                    'defaultValue' => $event->newValue['defaultValue'],
+                    'required'     => $event->newValue['required'],
+                    'handle'       => $event->newValue['handle'],
+                    'name'         => $event->newValue['name'],
+                    'desc'         => $event->newValue['desc']
+                ])->execute();
+        } else {
+            Craft::$app->db->createCommand()->update(self::TABLE, [
+                    'uid'          => $uid,
+                    'enabled'      => $event->newValue['enabled'],
+                    'defaultValue' => $event->newValue['defaultValue'],
+                    'required'     => $event->newValue['required'],
+                    'handle'       => $event->newValue['handle'],
+                    'name'         => $event->newValue['name'],
+                    'desc'         => $event->newValue['desc']
+                ], ['id' => $id])->execute();
+        }
+    }
+
+    public function handleDeleteConsentGroup(ConfigEvent $event)
+    {
+        $uid = $event->tokenMatches[0];
+        Craft::$app->db->createCommand()->delete(self::TABLE, ['uid' => $uid])->execute();
     }
 }